Split enrichment functions into their own file
